refactor: extract types and interfaces from constants into Types
