+++ frontend/src/components/layout/navbar/DesktopNav.tsx
@@ -1,6 +1,6 @@
 import { NavLink } from 'react-router-dom';
 
-import { type NavProps } from '@/constants/navigation';
+import { type NavProps } from '@/types/layout';
 
 const DesktopNav = ({ navItems }: NavProps) => {
   return (
