+++ frontend/src/constants/themes.ts
@@ -1,3 +1,3 @@
-export type Theme = 'light' | 'dark';
+import { type Theme } from "@/types/layout";
 
 export const DEFAULT_THEME: Theme = 'light';
