+++ frontend/src/types/layout.ts
@@ -1,0 +1,10 @@
+type NavItem = {
+  to: string;
+  label: string;
+};
+
+export interface NavProps {
+  navItems: NavItem[];
+}
+
+export type Theme = 'light' | 'dark';
